Move Reddit-related stuff into reddit.py module
